test: refactor test structure and add iac and app test commands to stack deployments
