+++ package.json
@@ -39,7 +39,7 @@
     "typescript-eslint": "^8.47.0"
   },
   "dependencies": {
-    "@orcabus/platform-cdk-constructs": "1.0.4",
+    "@orcabus/platform-cdk-constructs": "1.0.8",
     "aws-cdk-lib": "^2.227.0",
     "cargo-lambda-cdk": "^0.0.35",
     "constructs": "^10.4.3"
